Leverage the multicluster-runtime library to enable the operator to manage resources from many control planes.
A Datum provider is introduced that discovers Datum Projects and registers them
as clusters. This will eventually be moved out into a shared library.

Ginkgo test scaffolding has been removed, as we'll be leveraging Chainsaw.
Existing tests will be ported in a future iteration.

These changes depend on
https://github.com/multicluster-runtime/multicluster-runtime/pull/17, as well as
other changes to support kind clusters that do not have a prefix of "fleet-" in
the name that have not yet been published.