Remove completed TODOs, add additional debug output on failure, sleep before executing curl.
I plan to spend time and investigate a method to
determine if the Envoy replicas have received the
latest xDS config. With the way that Envoy Gateway
is written, it will update the k8s resources to be
accepted before the Envoy replicas have received
the new xDS config. In a resource constrained
environment, this delay may be large.