Programmed hostname retention, custom hostname test coverage.
Allow hostnames which have been successfully configured on the downstream
gateway stay on the gateway, regardless of whether or not there is a matching
Domain that's verified. This is done in case the user deletes the Domain record
after verification has completed. In the future, it would be better to place a
lien on the Domain via a ValidatingAdmissionPolicyBinding or higher level
abstraction. A finalizer is undesirable as many resources could leverage the
Domain, and we'd want to prohibit deletion, not just keep a deleted item from
going away.

For more thoughts on liens, the following issue provides good context:
https://github.com/kubernetes/kubernetes/issues/10179#issuecomment-2889238042

- Renamed the HostnamesReady condition to HostnamesVerified.
- Updated HTTPProxy status updates to only process the underlying Gateway's
  status if the Generation of the Accepted condition matches the Gateway's
  generation.