Default Gateway hostname changes to avoid issues outlined in GEP-3567.
Before these changes, downstream gateways were being provisioned with listeners
which are considered to have overlapping hostnames. This is a result of
configuring the `<uuid>.<targetDomain>` hostname on a listener, along with the
`[v4|v6].<uuid>.<targetDomain>` hostnames on listeners. When this occurs, Envoy
Gateway will disable HTTP2 on the listeners, requiring explicit ALPN
configuration via a ClientTrafficPolicy.

To make sure we provide users with gateways which support HTTP2 out of the box,
the decision was made to no longer configure the gateway with the v4 or v6
prefixed hostnames. These addresses will remain in DNS, responding only with
A or AAAA records.

In an effort to make it more clear to users which hostnames a gateway will
accept, the system now injects default HTTP and HTTPS listeners into gateways
created by the user which have the `<uuid>.<targetDomain>` hostname. The user
may remove these listeners if they wish. All three DNS addresses will remain in
the `Status.Addresses` field of the Gateway, which now makes this field more
aligned with the intent defined in the spec.

The default listener names are now `default-http` and `default-https`,
which will result in the reconfiguration of some gateway resources upon
deployment - particularly those maintained via HTTPProxy resources.

See: https://gateway-api.sigs.k8s.io/geps/gep-3567/