Support v1 BackendTLSPolicies from the v1.4.0-rc2 release of the Gateway API.
Considering BackendTLSPolicy is moving from v1alpha3 to v1 in the v1.4.0
release of the Gateway API, we'll expose that version for users to
integrate with. The v1alpha3 version is also available to use, and
will auto upgrade to v1. There are currently no structural changes between
the v1alpha3 and v1 versions of the CRD.

The operator will configure downstream BackendTLSPolicies with v1alpha3,
until Envoy Gateway is updated to support v1.

The only restriction added to BackendTLSPolicies is that they must target
Backend resources only.

The generic resource replication controller has been configured to replicate
BackendTLSPolicies between the upstream and downstream clusters, with a
tweak to update the upstream resource's status using the v1 version of
the CRD.