Fix bug with transitioning proxy target domain.
If an existing Gateway exists with a DNS address provisioned to
a previous target domain, it will attempt to verify ownership of the
previous domain, and will remove the hostname from being programmed in
the downstream gateway resource.

To address this, consider any value in the Gateway's Address of type
HostnameAddressType as verified.